requirements.txt
discord
requests
py3-validate-email
pymysql
flask
flask-limiter
waitress
